+++ docs/conf.py
@@ -78,7 +78,7 @@ master_doc = 'index'
 
 # General information about the project.
 project = 'AIRR Standards'
-copyright = '2017-2020, AIRR Community'
+copyright = '2015-2023, AIRR Community'
 author = 'AIRR Community'
 
 # The name of the Pygments (syntax highlighting) style to use.
@@ -113,7 +113,6 @@ htmlhelp_basename = 'airr-standardsdoc'
 # PyData options
 # html_theme = "pydata_sphinx_theme"
 html_theme = "sphinx_book_theme"
-html_logo = "_static/AIRR_logo-only.png"
 
 # Bootstrap options
 # html_theme = 'bootstrap'
@@ -122,7 +121,12 @@ html_logo = "_static/AIRR_logo-only.png"
 # html_sidebars = {'**': ['globaltoc.html']}
 # html_sidebars = {'**': ['globaltoc.html', 'sourcelink.html', 'searchbox.html']}
 # html_sidebars = {'**': ['searchbox.html', 'globaltoc.html']}
-# html_theme_options = {
+html_theme_options = {
+#     # Logo options replacing top-level "html_logo" option to facilitate light/dark theme switch)
+    'logo': {
+        'image_light': '_static/AIRR-C-Logo_outlined_light.svg',
+        'image_dark': '_static/AIRR-C-Logo_outlined_dark.svg'
+    }
 #     # Navigation bar title. (Default: ``project`` value)
 #     'navbar_title': 'AIRR Community Standards',
 #
@@ -189,7 +193,7 @@ html_logo = "_static/AIRR_logo-only.png"
 #     # Choose Bootstrap version.
 #     # Values: "3" (default) or "2" (in quotes)
 #     'bootstrap_version': '2',
-# }
+}
 
 # -- Options for LaTeX output ---------------------------------------------
 
@@ -318,13 +322,17 @@ def parse_schema(spec, schema):
                     base_dic = xairr['ontology']
                     ontology_format = (str(base_dic['top_node']['id']), str(base_dic['top_node']['label']) )
                     # Replace name with url-linked name
-                    data_format = 'Ontology: { top_node: { id: %s, value: %s}}' % (ontology_format)
+                    data_format = 'Ontology: { top_node: { id: %s, label: %s}}' % (ontology_format)
                     # Get 'type' for ontology
-                    example = 'id: %s, value: %s' % (example['id'], example['label'])
+                    example = 'id: %s, label: %s' % (example['id'], example['label'])
                 elif xairr['format'] == 'controlled vocabulary':
                     if attr.get('enum', None) is not None:
+                        if None in attr['enum']:
+                            attr['enum'].remove(None)
                         data_format = 'Controlled vocabulary: %s' % ', '.join(attr['enum'])
                     elif attr.get('items', None) is not None:
+                        if None in attr['items']['enum']:
+                            attr['items']['enum'].remove(None)
                         data_format = 'Controlled vocabulary: %s' % ', '.join(attr['items']['enum'])
         else:
             nullable = True
@@ -389,7 +397,7 @@ if not os.path.exists(download_path):  os.mkdir(download_path)
 # Write MiAIRR TSV
 fields = ['Set', 'Subset', 'Designation', 'Field', 'Type', 'Format', 'Level', 'Definition', 'Example']
 tables = ['Study', 'Subject', 'Diagnosis', 'Sample', 'CellProcessing', 'NucleicAcidProcessing',
-          'PCRTarget', 'SequencingRun', 'RawSequenceData', 'DataProcessing']
+          'PCRTarget', 'SequencingRun', 'SequencingData', 'DataProcessing']
 # tables = data_elements.keys()
 miairr_schema = []
 with open(os.path.join(download_path, '%s.tsv' % 'AIRR_Minimal_Standard_Data_Elements'), 'w') as f:
@@ -405,8 +413,10 @@ html_context['MiAIRR_schema'] = miairr_schema
 # Write individual spec TSVs
 fields = ['Name', 'Type', 'Attributes', 'Definition']
 tables = ['Repertoire', 'Study', 'Subject', 'Diagnosis', 'Sample', 'CellProcessing', 'NucleicAcidProcessing',
-          'PCRTarget', 'SequencingRun', 'RawSequenceData', 'DataProcessing',
-          'Rearrangement', 'Alignment', 'Clone', 'Tree', 'Node', 'Cell']
+          'PCRTarget', 'SequencingRun', 'SequencingData', 'DataProcessing',
+          'Rearrangement', 'Alignment', 'Clone', 'Tree', 'Node', 'Cell', 'CellExpression',
+          'RearrangedSequence', 'UnrearrangedSequence', 'SequenceDelineationV', 'AlleleDescription', 'GermlineSet',
+          'GenotypeSet', 'Genotype', 'MHCGenotypeSet', 'MHCGenotype', 'Receptor']
 for spec in tables:
     with open(os.path.join(download_path, '%s.tsv' % spec), 'w') as f:
         writer = csv.DictWriter(f, fieldnames=fields, dialect='excel-tab', extrasaction='ignore')