+++ lang/python/airr/interface.py
@@ -4,26 +4,28 @@ Interface functions for file operations
 from __future__ import absolute_import
 
 # System imports
+import gzip
+import json
 import sys
 import pandas as pd
-from collections import OrderedDict
-from itertools import chain
-from pkg_resources import resource_filename
-import json
 import yaml
 import yamlordereddictloader
+from collections import OrderedDict
+from itertools import chain
 from io import open
+from warnings import warn
 
 if (sys.version_info > (3, 0)):
     from io import StringIO
-else: # Python 2 code in this block
+else:
+    # Python 2 code in this block
     from io import BytesIO as StringIO
-
 
 # Load imports
 from airr.io import RearrangementReader, RearrangementWriter
-from airr.schema import ValidationError, RearrangementSchema, RepertoireSchema
+from airr.schema import Schema, RearrangementSchema, RepertoireSchema, AIRRSchema, DataFileSchema, ValidationError
 
+#### Rearrangement ####
 
 def read_rearrangement(filename, validate=False, debug=False):
     """
@@ -38,8 +40,12 @@ def read_rearrangement(filename, validate=False, debug=False):
     Returns:
       airr.io.RearrangementReader: iterable reader class.
     """
-
-    return RearrangementReader(open(filename, 'r'), validate=validate, debug=debug)
+    if filename.endswith(".gz"):
+        handle = gzip.open(filename, 'r')
+    else:
+        handle = open(filename, 'r')
+        
+    return RearrangementReader(handle, validate=validate, debug=debug)
 
 
 def create_rearrangement(filename, fields=None, debug=False):
@@ -216,6 +222,201 @@ def validate_rearrangement(filename, debug=False):
 
     return valid
 
+#### AIRR Data Model ####
+
+def read_airr(filename, format=None, validate=False, model=True, debug=False):
+    """
+    Load an AIRR Data file
+
+    Arguments:
+      filename (str): path to the input file.
+      format (str): input file format valid strings are "yaml" or "json". If set to None,
+                    the file format will be automatically detected from the file extension.
+      validate (bool): whether to validate data as it is read, raising a ValidationError
+                       exception in the event of a validation failure.
+      model (bool): If True only validate objects defined in the AIRR DataFile schema.
+                  If False, attempt validation of all top-level objects.
+                  Ignored if validate=False.
+      debug (bool): debug flag. If True print debugging information to standard error.
+
+    Returns:
+      dict: dictionary of AIRR Data objects.
+    """
+    # Because the AIRR Data File is read in completely, we do not bother with a reader class.
+    # Determine file type from extension and use appropriate loader
+    ext = str.lower(filename.split('.')[-1]) if not format else format
+    if ext in ('yaml', 'yml'):
+        with open(filename, 'r', encoding='utf-8') as handle:
+            data = yaml.load(handle, Loader=yamlordereddictloader.Loader)
+    elif ext == 'json':
+        with open(filename, 'r', encoding='utf-8') as handle:
+            data = json.load(handle)
+    else:
+        if debug:  sys.stderr.write('Unknown file type: %s. Supported file extensions are "yaml", "yml" or "json"\n' % ext)
+        raise TypeError('Unknown file type: %s. Supported file extensions are "yaml", "yml" or "json"\n' % ext)
+        data = None
+
+    # Validate if requested
+    if validate:
+        if debug:  sys.stderr.write('Validating: %s\n' % filename)
+        try:
+            valid = validate_airr(data, model=model, debug=debug)
+        except ValidationError as e:
+            if debug:  sys.stderr.write('%s failed validation\n' % filename)
+            raise ValidationError(e)
+
+    # We do not perform any additional processing
+    return data
+
+
+def validate_airr(data, model=True, debug=False):
+    """
+    Validates an AIRR Data file
+
+    Arguments:
+      data (dict): dictionary containing AIRR Data Model objects
+      model (bool): If True only validate objects defined in the AIRR DataFile schema.
+                  If False, attempt validation of all top-level objects
+      debug (bool): debug flag. If True print debugging information to standard error.
+
+    Returns:
+      bool: True if files passed validation, otherwise False.
+    """
+    # Type check that input type is either dict or OrderedDict
+    if not hasattr(data, 'items'):
+        if debug:  sys.stderr.write('Data parameter is not a dictionary\n')
+        raise TypeError('Data parameter is not a dictionary')
+
+    # Loop through each AIRR object and validate
+    valid = True
+    for k, object in data.items():
+        if k in ('Info', 'DataFile'):  continue
+        if not object:  continue
+
+        # Check for DataFile schema
+        if model and k not in DataFileSchema.properties:
+            if debug:  sys.stderr.write('Skipping non-DataFile object: %s\n' % k)
+            continue
+
+        # Get Schema
+        schema = AIRRSchema.get(k, Schema(k))
+
+        # Determine input type and set appropriate iterator
+        if hasattr(object, 'items'):
+            # Validate named array (dict)
+            obj_iter = object.items()
+            # Validate named array (dict) or a single object (dict)
+            # obj_iter = object.items() if 'definition' not in object.keys() else [0, object]
+        elif isinstance(object, list):
+            # Validate array
+            obj_iter = enumerate(object)
+        else:
+            # Unrecognized data structure
+            valid = False
+            if debug:  sys.stderr.write('%s is an unrecognized data structure: %s\n' % k)
+            continue
+
+        # Validate each record in array
+        for i, record in obj_iter:
+            try:
+                schema.validate_object(record)
+            except ValidationError as e:
+                valid = False
+                if debug:  sys.stderr.write('%s at array position %s with validation error: %s\n' % (k, i, e))
+
+    if not valid:
+        raise ValidationError('AIRR Data Model has validation failures')
+
+    return valid
+
+
+def write_airr(filename, data, format=None, info=None, validate=False, model=True, debug=False):
+    """
+    Write an AIRR Data file
+
+    Arguments:
+      filename (str): path to the output file.
+      data (dict): dictionary of AIRR Data Model objects.
+      format (str): output file format valid strings are "yaml" or "json". If set to None,
+                    the file format will be automatically detected from the file extension.
+      info (object): info object to write. Will write current AIRR Schema info if not specified.
+      validate (bool): whether to validate data before it is written, raising a ValidationError
+                       exception in the event of a validation failure.
+      model (bool): If True only validate and write objects defined in the AIRR DataFile schema.
+                  If False, attempt validation and write of all top-level objects
+      debug (bool): debug flag. If True print debugging information to standard error.
+
+    Returns:
+      bool: True if the file is written without error.
+    """
+    # Type check that input type is either dict or OrderedDict
+    if not hasattr(data, 'items'):
+        if debug:  sys.stderr.write('Data parameter is not a dictionary\n')
+        raise TypeError('Data parameter is not a dictionary')
+
+    # Validate if requested
+    if validate:
+        if debug:  sys.stderr.write('Validating: %s\n' % filename)
+        try:
+            valid = validate_airr(data, model=model, debug=debug)
+        except ValidationError as e:
+            if debug:  sys.stderr.write(e)
+            raise ValidationError(e)
+
+    md = OrderedDict()
+    if info is None:
+        info = RearrangementSchema.info.copy()
+        info['title'] = 'AIRR Data File'
+        info['description'] = 'AIRR Data File written by AIRR Standards Python Library'
+    md['Info'] = info
+
+    # Loop through each object and add them to the output dict
+    for k, obj in data.items():
+        if k in ('Info', 'DataFile'):  continue
+        if not obj:  continue
+        if model and k not in DataFileSchema.properties:
+            if debug:  sys.stderr.write('Skipping non-DataFile object: %s\n' % k)
+            continue
+        md[k] = obj
+
+    # Determine file type from extension and use appropriate loader
+    ext = str.lower(filename.split('.')[-1]) if not format else format
+    if ext in ('yaml', 'yml'):
+        with open(filename, 'w') as handle:
+            yaml.dump(md, handle, default_flow_style=False)
+    elif ext == 'json':
+        with open(filename, 'w') as handle:
+            json.dump(md, handle, sort_keys=False, indent=2)
+    else:
+        if debug:
+            sys.stderr.write('Unknown file type: %s. Supported file extensions are "yaml", "yml" or "json"\n' % ext)
+        raise TypeError('Unknown file type: %s. Supported file extensions are "yaml", "yml" or "json"\n' % ext)
+
+    return True
+
+
+#### Deprecated ####
+
+def repertoire_template():
+    """
+    Return a blank repertoire object from the template. This object has the complete
+    structure with all of the fields and all values set to None or empty string.
+
+    Returns:
+      object: empty repertoire object.
+
+    .. deprecated:: 1.4
+       Use :meth:`schema.Schema.template` instead.
+    """
+    # Deprecation
+    warn('repertoire_template is deprecated and will be removed in a future release.\nUse schema.Schema.template instead.\n',
+         DeprecationWarning, stacklevel=2)
+
+    # Build template
+    object = RepertoireSchema.template()
+
+    return object
+
 
 def load_repertoire(filename, validate=False, debug=False):
     """
@@ -228,24 +429,17 @@ def load_repertoire(filename, validate=False, debug=False):
       debug (bool): debug flag. If True print debugging information to standard error.
 
     Returns:
-      list: list of Repertoire dictionaries.
-    """
-    # Because the repertoires are read in completely, we do not bother
-    # with a reader class.
-    md = None
+      dict: dictionary of AIRR Data objects.
 
-    # determine file type from extension and use appropriate loader
-    ext = filename.split('.')[-1]
-    if ext in ('yaml', 'yml'):
-        with open(filename, 'r', encoding='utf-8') as handle:
-            md = yaml.load(handle, Loader=yamlordereddictloader.Loader)
-    elif ext == 'json':
-        with open(filename, 'r', encoding='utf-8') as handle:
-            md = json.load(handle)
-    else:
-        if debug:
-            sys.stderr.write('Unknown file type: %s. Supported file extensions are "yaml", "yml" or "json"\n' % (ext))
-        raise TypeError('Unknown file type: %s. Supported file extensions are "yaml", "yml" or "json"\n' % (ext))
+    .. deprecated:: 1.4
+       Use :func:`read_airr` instead.
+    """
+    # Deprecation
+    warn('load_repertoire is deprecated and will be removed in a future release.\nUse read_airr instead.\n',
+         DeprecationWarning, stacklevel=2)
+
+   # use standard load function, we only validate Repertoire if requested
+    md = read_airr(filename, validate=validate, debug=debug)
 
     if md.get('Repertoire') is None:
         if debug:
@@ -282,7 +476,14 @@ def validate_repertoire(filename, debug=False):
 
     Returns:
       bool: True if files passed validation, otherwise False.
+
+    .. deprecated:: 1.4
+       Use :func:`validate_airr` instead.
     """
+    # Deprecation
+    warn('validate_repertoire is deprecated and will be removed in a future release.\nUse validate_airr instead.\n',
+         DeprecationWarning, stacklevel=2)
+
     valid = True
     if debug:
         sys.stderr.write('Validating: %s\n' % filename)
@@ -313,7 +514,14 @@ def write_repertoire(filename, repertoires, info=None, debug=False):
 
     Returns:
       bool: True if the file is written without error.
+
+    .. deprecated:: 1.4
+       Use :func:`write_airr` instead.
     """
+    # Deprecation
+    warn('write_repertoire is deprecated and will be removed in a future release.\nUse write_airr instead.\n',
+         DeprecationWarning, stacklevel=2)
+
     if not isinstance(repertoires, list):
         if debug:
             sys.stderr.write('Repertoires parameter is not a list\n')
@@ -327,37 +535,4 @@ def write_repertoire(filename, repertoires, info=None, debug=False):
     md['Info'] = info
     md['Repertoire'] = repertoires
 
-    # determine file type from extension and use appropriate loader
-    ext = filename.split('.')[-1]
-    if ext == 'yaml' or ext == 'yml':
-        with open(filename, 'w') as handle:
-            md = yaml.dump(md, handle, default_flow_style=False)
-    elif ext == 'json':
-        with open(filename, 'w') as handle:
-            md = json.dump(md, handle, sort_keys=False, indent=2)
-    else:
-        if debug:
-            sys.stderr.write('Unknown file type: %s. Supported file extensions are "yaml", "yml" or "json"\n' % (ext))
-        raise TypeError('Unknown file type: %s. Supported file extensions are "yaml", "yml" or "json"\n' % (ext))
-        
-    return True
-
-
-def repertoire_template():
-    """
-    Return a blank repertoire object from the template. This object has the complete
-    structure with all of the fields and all values set to None or empty string.
-
-    Returns:
-      object: empty repertoire object.
-    """
-    
-    # TODO: I suppose we should dynamically create this from the schema
-    # versus loading a template.
-
-    # Load blank template
-    f = resource_filename(__name__, 'specs/blank.airr.yaml')
-    object = load_repertoire(f)
-
-    return object['Repertoire'][0]
-
+    return write_airr(filename, md, info=info, debug=debug)