+++ lang/python/airr/tools.py
@@ -20,6 +20,7 @@ AIRR tools and utilities
 # System imports
 import argparse
 import sys
+from warnings import warn
 
 # Local imports
 from airr import __version__
@@ -43,7 +44,7 @@ def merge_cmd(out_file, airr_files, drop=False, debug=False):
     return airr.interface.merge_rearrangement(out_file, airr_files, drop=drop, debug=debug)
 
 # internal wrapper function before calling validate interface method
-def validate_cmd(airr_files, debug=True):
+def validate_rearrangement_cmd(airr_files, debug=True):
     """
     Validates one or more AIRR rearrangements files
 
@@ -54,12 +55,46 @@ def validate_cmd(airr_files, debug=True):
     Returns:
       boolean: True if all files passed validation, otherwise False
     """
-    try:
-        valid = [airr.interface.validate_rearrangement(f, debug=debug) for f in airr_files]
-        return all(valid)
-    except Exception as err:
-        sys.stderr.write('Error occurred while validating AIRR rearrangement files: ' + str(err) + '\n')
-        return False
+    valid = []
+    for f in airr_files:
+        try:
+            v = airr.interface.validate_rearrangement(f, debug=debug)
+            valid.append(v)
+        except Exception as e:
+            sys.stderr.write('%s\n' % e)
+            sys.stderr.write('Validation failed for file: %s\n\n' % f)
+            valid.append(False)
+        else:
+            if not v:  sys.stderr.write('Validation failed for file: %s\n\n' % f)
+
+    return all(valid)
+
+def validate_airr_cmd(airr_files, debug=True):
+    """
+    Validates one or more AIRR Data Model files
+
+    Arguments:
+      airr_files (list): list of input files to validate.
+      debug (bool): debug flag. If True print debugging information to standard error.
+
+    Returns:
+      boolean: True if all files passed validation, otherwise False
+    """
+    valid = []
+    for f in airr_files:
+        if debug: sys.stderr.write('Validating: %s\n' % f)
+        try:
+            data = airr.interface.read_airr(f, validate=False, debug=debug)
+            v = airr.interface.validate_airr(data, debug=debug)
+            valid.append(v)
+        except Exception as e:
+            sys.stderr.write('%s\n' % e)
+            sys.stderr.write('Validation failed for file: %s\n\n' % f)
+            valid.append(False)
+
+    return all(valid)
+
+#### Deprecated ####
 
 # internal wrapper function before calling validate interface method
 def validate_repertoire_cmd(airr_files, debug=True):
@@ -73,12 +108,21 @@ def validate_repertoire_cmd(airr_files, debug=True):
     Returns:
       boolean: True if all files passed validation, otherwise False
     """
-    try:
-        valid = [airr.interface.validate_repertoire(f, debug=debug) for f in airr_files]
-        return all(valid)
-    except Exception as err:
-        sys.stderr.write('Error occurred while validating AIRR repertoire metadata files: ' + str(err) + '\n')
-        return False
+    # Deprecation
+    warn('validate_repertoire_cmd is deprecated and will be removed in a future release.\nUse =validate_airr_cmd instead.\n',
+         DeprecationWarning, stacklevel=2)
+
+    valid = []
+    for f in airr_files:
+        try:
+            v = airr.interface.validate_repertoire(f, debug=debug)
+            valid.append(v)
+        except Exception as e:
+            sys.stderr.write('%s\n' % e)
+            sys.stderr.write('Validation failed for file: %s\n\n' % f)
+            valid.append(False)
+
+    return all(valid)
 
 def define_args():
     """
@@ -137,20 +181,10 @@ def define_args():
     # Subparser to validate files
     parser_validate = subparsers.add_parser('validate', parents=[common_parser],
                                             add_help=False,
-                                            help='Validate AIRR files.',
-                                            description='Validate AIRR files.')
+                                            help='Validate files for AIRR Standards compliance.',
+                                            description='Validate files for AIRR Standards compliance.')
     validate_subparser = parser_validate.add_subparsers(title='subcommands', metavar='',
                                        help='Database operation')
-
-    # Subparser to validate repertoire files
-    parser_validate = validate_subparser.add_parser('repertoire', parents=[common_parser],
-                                            add_help=False,
-                                            help='Validate AIRR repertoire metadata files.',
-                                            description='Validate AIRR repertoire metadata files.')
-    group_validate = parser_validate.add_argument_group('validate arguments')
-    group_validate.add_argument('-a', nargs='+', action='store', dest='airr_files', required=True,
-                                help='A list of AIRR repertoire metadata files.')
-    parser_validate.set_defaults(func=validate_repertoire_cmd)
 
     # Subparser to validate rearrangement files
     parser_validate = validate_subparser.add_parser('rearrangement', parents=[common_parser],
@@ -160,7 +194,27 @@ def define_args():
     group_validate = parser_validate.add_argument_group('validate arguments')
     group_validate.add_argument('-a', nargs='+', action='store', dest='airr_files', required=True,
                                 help='A list of AIRR rearrangement files.')
-    parser_validate.set_defaults(func=validate_cmd)
+    parser_validate.set_defaults(func=validate_rearrangement_cmd)
+
+    # Subparser to validate AIRR Data Model files
+    parser_validate = validate_subparser.add_parser('airr', parents=[common_parser],
+                                            add_help=False,
+                                            help='Validate AIRR Data Model files.',
+                                            description='Validate AIRR Data Model files.')
+    group_validate = parser_validate.add_argument_group('validate arguments')
+    group_validate.add_argument('-a', nargs='+', action='store', dest='airr_files', required=True,
+                                help='A list of AIRR Data Model files.')
+    parser_validate.set_defaults(func=validate_airr_cmd)
+
+    # Subparser to validate repertoire files
+    parser_validate = validate_subparser.add_parser('repertoire', parents=[common_parser],
+                                                    add_help=False,
+                                                    help='Validate AIRR repertoire metadata files.',
+                                                    description='Validate AIRR repertoire metadata files.')
+    group_validate = parser_validate.add_argument_group('validate arguments')
+    group_validate.add_argument('-a', nargs='+', action='store', dest='airr_files', required=True,
+                                help='A list of AIRR repertoire metadata files.')
+    parser_validate.set_defaults(func=validate_repertoire_cmd)
 
     return parser
 
@@ -180,6 +234,11 @@ def main():
     args_dict = args.__dict__.copy()
     del args_dict['command']
     del args_dict['func']
+
+    # Deprecation warnings
+    if args.func is validate_repertoire_cmd:
+        print('The "validate repertoire" subcommand is deprecated and will be removed in a future release.',
+              '\nUse the "validate airr" subcommand instead.\n')
 
     # Call tool function
     result = args.func(**args_dict)